+++ delta-coverage-gradle/src/functionalTest/resources/multi-module-test-project/module2/src/intTest/java/com/module2/Class2Test.java
@@ -5,9 +5,6 @@ import org.junit.jupiter.api.Test;
 public class Class2Test {
     @Test
     public void printMin() {
-        new Class2().min(1, 2);
         new Class2().min(2, 1);
-
-        new Class2().max(1, 2);
     }
 }