+++ delta-coverage-gradle/src/main/kotlin/io/github/surpsg/deltacoverage/gradle/task/NativeGitDiffTask.kt
@@ -1,4 +1,4 @@
-package io.github.surpsg.deltacoverage.gradle
+package io.github.surpsg.deltacoverage.gradle.task
 
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.RegularFileProperty